+++ crates/core/src/contract/executor/runtime.rs
@@ -1,11 +1,15 @@
+mod contract_ops;
+mod delegates;
 mod executor_impl;
 mod pool;
+mod subscriptions;
 
 use super::*;
 use super::{
     ContractExecutor, ContractRequest, ContractResponse, ExecutorError, InitCheckResult,
     RequestError, Response, SLOW_INIT_THRESHOLD, STALE_INIT_THRESHOLD, StateStoreError, now_nanos,
 };
+pub(crate) use contract_ops::ReclaimOutcome;
 pub use pool::RuntimePool;
 
 /// Maximum number of related contracts a single validation can request.
@@ -565,876 +569,6 @@ impl Executor<Runtime> {
 
         result
     }
-
-    /// Export this hosted user's per-user delegate secrets into an encrypted
-    /// bundle, sealed under the user's `token` (hosted-mode export, P3-live of
-    /// #4381).
-    ///
-    /// Runs entirely on the executor (which owns the `SecretsStore` via its
-    /// `Runtime`), so the on-disk redb is touched only by its single writer.
-    /// The bundle is scoped to `user_context.scope()` — strictly the per-user
-    /// namespace, never `Local`. The `token` is the bundle-key material so the
-    /// user re-imports on their own peer with the token they already hold.
-    ///
-    /// The token and the derived key material live only in borrowed/`Zeroizing`
-    /// buffers here and inside `export_secret_bundle`; nothing is logged.
-    pub fn export_user_secrets(
-        &self,
-        user_context: &UserSecretContext,
-        token: &[u8],
-    ) -> Result<Vec<u8>, ExecutorError> {
-        use crate::wasm_runtime::secret_export::BundleKeyMaterial;
-        self.runtime
-            .export_secret_bundle(user_context.scope(), &BundleKeyMaterial::Token(token))
-            .map_err(|e| ExecutorError::other(anyhow::anyhow!("secret export failed: {e}")))
-    }
-
-    pub fn delegate_request(
-        &mut self,
-        req: DelegateRequest<'_>,
-        origin_contract: Option<&ContractInstanceId>,
-        caller_delegate: Option<&DelegateKey>,
-        user_context: Option<&UserSecretContext>,
-    ) -> Response {
-        // Mutual exclusion invariant: a single inbound delegate request is
-        // either dispatched on behalf of a contract-backed web app
-        // (`origin_contract = Some`) or on behalf of another delegate
-        // (`caller_delegate = Some`), never both. The doc comment on
-        // `ContractExecutor::execute_delegate_request` states this. The
-        // `debug_assert!` turns the convention into a tripwire so a future
-        // call site that violates it fails loudly in debug/test builds; in
-        // release builds the precedence below silently picks `caller_delegate`
-        // (fail-safe in the direction of "least surprising attestation").
-        debug_assert!(
-            !(origin_contract.is_some() && caller_delegate.is_some()),
-            "execute_delegate_request: at most one of origin_contract and \
-             caller_delegate may be Some (got both)"
-        );
-        tracing::debug!(
-            origin_contract = ?origin_contract,
-            caller_delegate = ?caller_delegate.map(|k| k.to_string()),
-            "received delegate request"
-        );
-        match req {
-            DelegateRequest::RegisterDelegate {
-                delegate,
-                cipher,
-                nonce,
-            } => {
-                use chacha20poly1305::{KeyInit, XChaCha20Poly1305};
-                let key = delegate.key().clone();
-                let arr = (&cipher).into();
-                let cipher = XChaCha20Poly1305::new(arr);
-                let nonce = nonce.into();
-                if let Some(contract) = origin_contract {
-                    self.delegate_origin_ids
-                        .entry(key.clone())
-                        .or_default()
-                        .push(*contract);
-                }
-                match self.runtime.register_delegate(delegate, cipher, nonce) {
-                    Ok(_) => Ok(DelegateResponse {
-                        key,
-                        values: Vec::new(),
-                    }),
-                    Err(err) => {
-                        tracing::warn!(
-                            delegate_key = %key,
-                            error = %err,
-                            phase = "register_failed",
-                            "Failed to register delegate"
-                        );
-                        Err(ExecutorError::other(StdDelegateError::RegisterError(key)))
-                    }
-                }
-            }
-            DelegateRequest::UnregisterDelegate(key) => {
-                self.delegate_origin_ids.remove(&key);
-
-                // Remove delegate from all contract subscription entries
-                crate::wasm_runtime::DELEGATE_SUBSCRIPTIONS.retain(|_, subscribers| {
-                    subscribers.remove(&key);
-                    !subscribers.is_empty()
-                });
-
-                // Clean up delegate creation tracking to prevent unbounded growth
-                crate::wasm_runtime::DELEGATE_INHERITED_ORIGINS.remove(&key);
-
-                // Decrement the global created-delegates counter so the slot can be reused.
-                // Only decrement if count > 0 to avoid underflow for delegates not created
-                // via the host function (e.g., registered directly by apps).
-                {
-                    use std::sync::atomic::Ordering;
-                    let count = &crate::wasm_runtime::CREATED_DELEGATES_COUNT;
-                    let prev = count.load(Ordering::Relaxed);
-                    if prev > 0 {
-                        count.fetch_sub(1, Ordering::Relaxed);
-                    }
-                }
-
-                match self.runtime.unregister_delegate(&key) {
-                    Ok(_) => Ok(HostResponse::Ok),
-                    Err(err) => {
-                        tracing::warn!(
-                            delegate_key = %key,
-                            error = %err,
-                            phase = "unregister_failed",
-                            "Failed to unregister delegate"
-                        );
-                        Ok(HostResponse::Ok)
-                    }
-                }
-            }
-            DelegateRequest::ApplicationMessages {
-                key,
-                inbound,
-                params,
-            } => {
-                let origin = resolve_message_origin(caller_delegate, origin_contract, &key);
-                match self.runtime.inbound_app_message(
-                    &key,
-                    &params,
-                    origin.as_ref(),
-                    // The per-user secret scope, present only in hosted mode and
-                    // derived solely from the connection token. It is delivered
-                    // here on a SEPARATE channel from `origin`/the request body,
-                    // so neither WASM nor any delegate-message content can set or
-                    // change which user's namespace a secret op touches.
-                    user_context,
-                    inbound
-                        .into_iter()
-                        .map(InboundDelegateMsg::into_owned)
-                        .collect(),
-                ) {
-                    Ok(values) => Ok(DelegateResponse { key, values }),
-                    Err(err) => {
-                        let key_display = key.to_string();
-                        let exec_err =
-                            ExecutorError::execution(err, Some(InnerOpError::Delegate(key)));
-                        // Downgrade "not found" to warn — expected during legacy
-                        // migration probes when old delegate WASM isn't on this node
-                        if exec_err.is_missing_delegate() {
-                            tracing::warn!(
-                                delegate_key = %key_display,
-                                "Delegate not found in store (expected for migration probes)"
-                            );
-                        } else {
-                            tracing::error!(
-                                delegate_key = %key_display,
-                                error = %exec_err,
-                                phase = "execution_failed",
-                                "Failed executing delegate"
-                            );
-                        }
-                        Err(exec_err)
-                    }
-                }
-            }
-            _ => Err(ExecutorError::other(anyhow::anyhow!("not supported"))),
-        }
-    }
-
-    async fn perform_contract_put(
-        &mut self,
-        contract: ContractContainer,
-        state: WrappedState,
-        related_contracts: RelatedContracts<'_>,
-    ) -> Response {
-        let key = contract.key();
-        let params = contract.params();
-
-        if self.get_local_contract(key.id()).await.is_ok() {
-            // Contract already exists — merge states locally and broadcast async.
-            //
-            // We intentionally do NOT delegate to perform_contract_update here because
-            // its network mode path uses op_request() which blocks waiting for the
-            // network operation to complete (120s timeout). For client-initiated puts
-            // (e.g. fdev publish), the client needs a timely response. The network
-            // broadcast is fire-and-forget — if it fails, subscribers will still get
-            // the update via their next sync.
-            //
-            // NOTE: This simplified path does not handle contracts that require related
-            // contracts for update_state or validate_state. If update_state returns
-            // MissingRelated (new_state=None with non-empty related), it is treated as
-            // "no change". This is acceptable because no current contracts use related
-            // contracts (see issue #2870 for completing that mechanism).
-            let current_state = self
-                .state_store
-                .get(&key)
-                .await
-                .map_err(ExecutorError::other)?
-                .clone();
-
-            let update = UpdateData::State(state.into());
-            let update_result = self
-                .runtime
-                .update_state(&key, &params, &current_state, &[update])
-                .map_err(|err| ExecutorError::execution(err, Some(InnerOpError::Upsert(key))))?;
-
-            // If update_state produced no new state, or the merged state is identical
-            // to current, return early with the current summary (no work to persist).
-            let new_state = match update_result.new_state {
-                Some(s) if s.as_ref() != current_state.as_ref() => {
-                    WrappedState::new(s.into_bytes())
-                }
-                _ => {
-                    let summary = self
-                        .runtime
-                        .summarize_state(&key, &params, &current_state)
-                        .map_err(|e| ExecutorError::execution(e, None))?;
-                    return Ok(ContractResponse::UpdateResponse { key, summary }.into());
-                }
-            };
-
-            // Validate before persisting (fetch related contracts from network if needed)
-            let validate_result = self
-                .fetch_related_for_validation_network(&key, &params, &new_state, &related_contracts)
-                .await?;
-            if validate_result != ValidateResult::Valid {
-                return Err(Self::validation_error_put(key, validate_result));
-            }
-
-            // Commit locally
-            let written_bytes = new_state.as_ref().len();
-            self.state_store
-                .update(&key, new_state.clone())
-                .await
-                .map_err(ExecutorError::other)?;
-            // State-write chokepoint (re-PUT): delegate to
-            // `Ring::commit_state_write` for bump + refresh + report. See
-            // its rustdoc and `RuntimePool::remove_contract` for the
-            // EvictContract re-host race this closes.
-            if let Some(op_manager) = &self.op_manager {
-                op_manager.ring.commit_state_write(&key, written_bytes);
-            }
-
-            self.send_update_notification(&key, &params, &new_state)
-                .await
-                .map_err(|_| {
-                    ExecutorError::request(StdContractError::Put {
-                        key,
-                        cause: "failed while sending notifications".into(),
-                    })
-                })?;
-
-            self.broadcast_state_change(key, new_state.clone()).await;
-
-            let summary = self
-                .runtime
-                .summarize_state(&key, &params, &new_state)
-                .map_err(|e| ExecutorError::execution(e, None))?;
-            return Ok(ContractResponse::UpdateResponse { key, summary }.into());
-        }
-
-        self.verify_and_store_contract(state.clone(), contract, related_contracts)
-            .await?;
-
-        self.send_update_notification(&key, &params, &state)
-            .await
-            .map_err(|_| {
-                ExecutorError::request(StdContractError::Put {
-                    key,
-                    cause: "failed while sending notifications".into(),
-                })
-            })?;
-
-        self.broadcast_state_change(key, state.clone()).await;
-
-        Ok(ContractResponse::PutResponse { key }.into())
-    }
-
-    async fn perform_contract_update(
-        &mut self,
-        key: ContractKey,
-        update: UpdateData<'_>,
-    ) -> Response {
-        let parameters = {
-            self.state_store
-                .get_params(&key)
-                .await
-                .map_err(ExecutorError::other)?
-                .ok_or_else(|| {
-                    RequestError::ContractError(StdContractError::Update {
-                        cause: "missing contract parameters".into(),
-                        key,
-                    })
-                })?
-        };
-
-        let current_state = self
-            .state_store
-            .get(&key)
-            .await
-            .map_err(ExecutorError::other)?;
-
-        let updates = vec![update];
-
-        // `Executor::contract_requests` is only invoked from `run_local_node`
-        // (HTTP/WS local-only entry points). Network-mode UPDATEs from
-        // clients arrive through `client_event_handling` →
-        // `start_client_update` and never reach this function.
-        if self.mode == OperationMode::Local {
-            let new_state = self
-                .get_updated_state(&parameters, current_state, key, updates)
-                .await?;
-            let summary = self
-                .runtime
-                .summarize_state(&key, &parameters, &new_state)
-                .map_err(|e| ExecutorError::execution(e, None))?;
-            return Ok(ContractResponse::UpdateResponse { key, summary }.into());
-        }
-
-        Err(ExecutorError::other(anyhow::anyhow!(
-            "network UPDATE must dispatch via `start_client_update` (client_events.rs); \
-             `perform_contract_update` is reachable only in local mode"
-        )))
-    }
-
-    /// Given a contract and a series of delta updates, it will try to perform an update
-    /// to the contract state and return the new state. If it fails to update the state,
-    /// it will return an error.
-    ///
-    /// If there are missing updates for related contracts, it will try to fetch them from the network.
-    async fn get_updated_state(
-        &mut self,
-        parameters: &Parameters<'_>,
-        current_state: WrappedState,
-        key: ContractKey,
-        mut updates: Vec<UpdateData<'_>>,
-    ) -> Result<WrappedState, ExecutorError> {
-        let new_state = {
-            let start = Instant::now();
-            loop {
-                let state_update_res = self
-                    .attempt_state_update(parameters, &current_state, &key, &updates)
-                    .await?;
-                let missing = match state_update_res {
-                    Either::Left(new_state) => {
-                        break new_state;
-                    }
-                    Either::Right(missing) => missing,
-                };
-                // some required contracts are missing
-                let required_contracts = missing.len() + 1;
-                for RelatedContract {
-                    contract_instance_id: id,
-                    mode,
-                } in missing
-                {
-                    // Try to look up the full key; if not found, treat as missing
-                    let local_state = if let Some(related_key) = self.lookup_key(&id) {
-                        self.state_store.get(&related_key).await.ok()
-                    } else {
-                        None
-                    };
-
-                    match local_state {
-                        Some(state) => {
-                            // in this case we are already subscribed to and are updating this contract,
-                            // we can try first with the existing value
-                            updates.push(UpdateData::RelatedState {
-                                related_to: id,
-                                state: state.into(),
-                            });
-                        }
-                        None => {
-                            let state = match self.local_state_or_from_network(&id, false).await? {
-                                Either::Left(state) => state,
-                                Either::Right(GetResult {
-                                    state, contract, ..
-                                }) => {
-                                    let Some(contract) = contract else {
-                                        return Err(ExecutorError::request(
-                                            RequestError::ContractError(StdContractError::Get {
-                                                key,
-                                                cause: "Missing contract".into(),
-                                            }),
-                                        ));
-                                    };
-                                    self.verify_and_store_contract(
-                                        state.clone(),
-                                        contract.clone(),
-                                        RelatedContracts::default(),
-                                    )
-                                    .await?;
-                                    state
-                                }
-                            };
-                            updates.push(UpdateData::State(state.into()));
-                            match mode {
-                                RelatedMode::StateOnce => {}
-                                RelatedMode::StateThenSubscribe => {
-                                    // After storing, we should be able to look up the key
-                                    if let Some(related_key) = self.lookup_key(&id) {
-                                        self.subscribe(related_key).await?;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                if updates.len() + 1 /* includes the original contract being updated update */ >= required_contracts
-                {
-                    // try running again with all the related contracts retrieved
-                    continue;
-                } else if start.elapsed() > Duration::from_secs(10) {
-                    /* make this timeout configurable, and anyway should be controlled globally*/
-                    return Err(RequestError::Timeout.into());
-                }
-            }
-        };
-
-        // Validate before persisting or broadcasting (fetch related contracts from network if needed).
-        let result = self
-            .fetch_related_for_validation_network(
-                &key,
-                parameters,
-                &new_state,
-                &RelatedContracts::default(),
-            )
-            .await?;
-
-        if result != ValidateResult::Valid {
-            return Err(Self::validation_error(key, result));
-        }
-        if new_state.as_ref() != current_state.as_ref() {
-            self.commit_state_update(&key, parameters, &new_state)
-                .await?;
-        }
-        Ok(new_state)
-    }
-
-    async fn get_local_contract(
-        &self,
-        id: &ContractInstanceId,
-    ) -> Result<State<'static>, Either<Box<RequestError>, anyhow::Error>> {
-        let Some(full_key) = self.lookup_key(id) else {
-            return Err(Either::Right(
-                StdContractError::MissingRelated { key: *id }.into(),
-            ));
-        };
-        let Ok(contract) = self.state_store.get(&full_key).await else {
-            return Err(Either::Right(
-                StdContractError::MissingRelated { key: *id }.into(),
-            ));
-        };
-        // SAFETY: `contract` is alive for the remainder of this function,
-        // and `state` does not escape this scope, so the reborrowed slice
-        // remains valid for its entire use.
-        let state: &[u8] = unsafe { std::mem::transmute::<&[u8], &'_ [u8]>(contract.as_ref()) };
-        Ok(State::from(state))
-    }
-
-    /// Verify and store a contract with depth=1 related contract resolution.
-    ///
-    /// 1. Store the contract code in the runtime store
-    /// 2. Validate state (fetching related contracts if requested, one round only)
-    /// 3. If valid, persist to state_store
-    async fn verify_and_store_contract(
-        &mut self,
-        state: WrappedState,
-        contract: ContractContainer,
-        related_contracts: RelatedContracts<'_>,
-    ) -> Result<(), ExecutorError> {
-        let key = contract.key();
-        let params = contract.params();
-        let state_hash = blake3::hash(state.as_ref());
-
-        tracing::debug!(
-            contract = %key,
-            state_size = state.as_ref().len(),
-            state_hash = %state_hash,
-            params_size = params.as_ref().len(),
-            "starting contract verification and storage"
-        );
-
-        // Store contract code in runtime store
-        self.runtime
-            .contract_store
-            .store_contract(contract)
-            .map_err(|e| {
-                tracing::error!(
-                    contract = %key,
-                    error = %e,
-                    "failed to store contract in runtime"
-                );
-                ExecutorError::other(e)
-            })?;
-
-        // Validate with depth=1 related contract resolution.
-        //
-        // DEPTH PROTECTION: fetch_related_for_validation enforces depth=1 —
-        // if validate_state returns RequestRelated(ids), we fetch those contracts
-        // and retry exactly once. A second RequestRelated is rejected as an error.
-        // This prevents:
-        //   - Recursive depth (related contracts requesting their own related contracts)
-        //   - Amplification attacks (contract requesting new contracts on every retry)
-        //   - Self-reference (contract requesting its own state)
-        //   - Excessive fan-out (max 10 related contracts per request)
-        // See MAX_RELATED_CONTRACTS_PER_REQUEST and RELATED_FETCH_TIMEOUT constants.
-        let result = self
-            .fetch_related_for_validation_network(&key, &params, &state, &related_contracts)
-            .await
-            .inspect_err(|_| {
-                if let Err(e) = self.runtime.contract_store.remove_contract(&key) {
-                    tracing::warn!(contract = %key, error = %e, "failed to remove contract after validation failure");
-                }
-            })?;
-
-        // fetch_related_for_validation resolves RequestRelated internally,
-        // so only Valid or Invalid are possible here.
-        if result != ValidateResult::Valid {
-            if let Err(e) = self.runtime.contract_store.remove_contract(&key) {
-                tracing::warn!(contract = %key, error = %e, "failed to remove contract after invalid validation");
-            }
-            return Err(ExecutorError::request(StdContractError::Put {
-                key,
-                cause: "not valid".into(),
-            }));
-        }
-
-        tracing::debug!(
-            contract = %key,
-            state_size = state.as_ref().len(),
-            "storing contract state"
-        );
-        let written_bytes = state.as_ref().len();
-        self.state_store
-            .store(key, state, params)
-            .await
-            .map_err(|e| {
-                tracing::error!(
-                    contract = %key,
-                    error = %e,
-                    "failed to store contract state"
-                );
-                ExecutorError::other(e)
-            })?;
-        // State-write chokepoint (verify_and_store PUT): delegate to
-        // `Ring::commit_state_write` for bump + refresh + report. See
-        // its rustdoc and `RuntimePool::remove_contract` for the
-        // EvictContract re-host race this closes.
-        if let Some(op_manager) = &self.op_manager {
-            op_manager.ring.commit_state_write(&key, written_bytes);
-        }
-
-        Ok(())
-    }
-
-    /// Reclaim a contract's on-disk storage after it was evicted from the
-    /// hosting cache.
-    ///
-    /// Deletes (1) the persisted state and parameters from the `StateStore`
-    /// and (2) the WASM code blob from the `ContractStore`. The contract-store
-    /// removal is code-hash refcount-safe: the shared `.wasm` blob is only
-    /// deleted once no other contract instance references the same code.
-    ///
-    /// Both steps are best-effort and independent: if one fails, the other is
-    /// still attempted so a partial reclaim is achieved rather than none. The
-    /// method is idempotent — both `StateStore::delete` and
-    /// `ContractStore::remove_contract` tolerate already-missing entries — so a
-    /// double eviction is harmless.
-    ///
-    /// Return value:
-    ///   - `Ok(ReclaimOutcome::Full)` — both halves are absent at end (either
-    ///     both deleted in this call, or one was already missing and the other
-    ///     was deleted, or both were already missing).
-    ///   - `Ok(ReclaimOutcome::Partial)` — exactly one half failed with a real
-    ///     error while the other succeeded. The caller MUST retain the
-    ///     pending-reclamation entry so a future sweep retries the remaining
-    ///     work. Closes the disk-leak edge case where a transient DB/FS
-    ///     error in one half leaves the other half permanently leaked. See
-    ///     PR #4212 review round 8.
-    ///   - `Err` — BOTH halves failed; surfaced so the caller can log/retry.
-    ///
-    /// This is the inherent implementation; the `ContractExecutor::remove_contract`
-    /// trait method delegates to it and translates the outcome into
-    /// pending-reclamation management.
-    async fn reclaim_contract_storage(
-        &mut self,
-        key: &ContractKey,
-    ) -> Result<ReclaimOutcome, ExecutorError> {
-        let state_result = match self.state_store.delete(key).await {
-            Ok(()) => Ok(()),
-            Err(e) => {
-                tracing::warn!(
-                    contract = %key,
-                    error = %e,
-                    "failed to delete persisted state while reclaiming evicted contract"
-                );
-                Err(())
-            }
-        };
-        let code_result = match self.runtime.contract_store.remove_contract(key) {
-            Ok(()) => Ok(()),
-            Err(e) => {
-                tracing::warn!(
-                    contract = %key,
-                    error = %e,
-                    "failed to delete WASM code while reclaiming evicted contract"
-                );
-                Err(())
-            }
-        };
-
-        let state_ok = state_result.is_ok();
-        let code_ok = code_result.is_ok();
-        if !state_ok && !code_ok {
-            return Err(ExecutorError::other(anyhow::anyhow!(
-                "failed to reclaim any on-disk storage for contract {key}"
-            )));
-        }
-
-        let outcome = if state_ok && code_ok {
-            ReclaimOutcome::Full
-        } else {
-            ReclaimOutcome::Partial
-        };
-        tracing::info!(
-            contract = %key,
-            state_deleted = state_ok,
-            code_deleted = code_ok,
-            ?outcome,
-            "reclaimed on-disk storage for evicted contract"
-        );
-        Ok(outcome)
-    }
-}
-
-/// Outcome of [`Executor::reclaim_contract_storage`].
-///
-/// The split exists so the caller (`RuntimePool::remove_contract`) can decide
-/// whether to clear the pending-reclamation entry (on `Full`) or leave it for
-/// a future retry (on `Partial`). See PR #4212 review round 8.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum ReclaimOutcome {
-    /// Both state and code are absent at end of the reclaim call.
-    Full,
-    /// Exactly one half failed with a real error; the other succeeded (or was
-    /// already absent). The pending-reclamation entry should be retained so a
-    /// future sweep retries the remaining work.
-    Partial,
-}
-
-impl Executor<Runtime> {
-    /// Network-aware variant retained for the PUT path.
-    ///
-    /// The base `fetch_related_for_validation` on the bridged impl now also
-    /// escalates to network GET via `op_manager` when the local state_store
-    /// lookup misses, so the two implementations are functionally equivalent
-    /// for `Executor<Runtime>`. This variant is kept because it exposes the
-    /// `local_state_or_from_network` helper directly and several PUT call
-    /// sites already wire through it; collapsing the two would touch more
-    /// surface area than the bug fix needs.
-    async fn fetch_related_for_validation_network(
-        &mut self,
-        key: &ContractKey,
-        params: &Parameters<'_>,
-        state: &WrappedState,
-        initial_related: &RelatedContracts<'_>,
-    ) -> Result<ValidateResult, ExecutorError> {
-        let result = self
-            .runtime
-            .validate_state(key, params, state, initial_related)
-            .map_err(|e| ExecutorError::execution(e, None))?;
-
-        let requested_ids = match result {
-            ValidateResult::Valid | ValidateResult::Invalid => return Ok(result),
-            ValidateResult::RequestRelated(ids) => ids,
-        };
-
-        // Apply the same safety checks as the base helper
-        if requested_ids.is_empty() {
-            return Err(ExecutorError::request(StdContractError::Put {
-                key: *key,
-                cause: "contract requested related contracts but provided empty list".into(),
-            }));
-        }
-        let self_id = key.id();
-        if requested_ids.iter().any(|id| id == self_id) {
-            return Err(ExecutorError::request(StdContractError::Put {
-                key: *key,
-                cause: "contract cannot request itself as a related contract".into(),
-            }));
-        }
-        let unique_ids: HashSet<ContractInstanceId> = requested_ids.into_iter().collect();
-        if unique_ids.len() > MAX_RELATED_CONTRACTS_PER_REQUEST {
-            return Err(ExecutorError::request(StdContractError::Put {
-                key: *key,
-                cause: format!(
-                    "contract requested {} related contracts, limit is {}",
-                    unique_ids.len(),
-                    MAX_RELATED_CONTRACTS_PER_REQUEST
-                )
-                .into(),
-            }));
-        }
-
-        tracing::debug!(
-            contract = %key,
-            related_count = unique_ids.len(),
-            "Fetching related contracts (with network fallback) for validation"
-        );
-
-        let mut related_map: HashMap<ContractInstanceId, Option<State<'static>>> =
-            HashMap::with_capacity(unique_ids.len());
-
-        // Parallel fetch — see fetch_related_for_validation for rationale
-        // (freenet/freenet-core#4077). The serial loop here had the same
-        // 10s/N effective per fetch problem.
-        //
-        // We can't reuse the `&mut self` `local_state_or_from_network`
-        // helper across multiple concurrent futures, so the per-id body
-        // is inlined: try the local state_store first, escalate to
-        // `fetch_related_via_network` (which only borrows
-        // `&Option<Arc<OpManager>>`). Reborrow as `&Self` so the
-        // per-id futures share an immutable borrow; the outer
-        // `&mut self` is reclaimed once `fetch_all` is awaited.
-        let this: &Self = &*self;
-        let fetch_all = async {
-            let results: Vec<(ContractInstanceId, Result<State<'static>, ExecutorError>)> =
-                futures::future::join_all(unique_ids.iter().map(|id| {
-                    let id = *id;
-                    async move {
-                        if let Some(full_key) = this.lookup_key(&id) {
-                            if let Ok(state) = this.state_store.get(&full_key).await {
-                                return (id, Ok(State::from(state.as_ref().to_vec())));
-                            }
-                        }
-                        let outcome = fetch_related_via_network(this.op_manager.as_ref(), &id)
-                            .await
-                            .map(|state| State::from(state.as_ref().to_vec()));
-                        (id, outcome)
-                    }
-                }))
-                .await;
-            for (id, res) in results {
-                related_map.insert(id, Some(res?));
-            }
-            Ok::<(), ExecutorError>(())
-        };
-
-        match tokio::time::timeout(RELATED_FETCH_TIMEOUT, fetch_all).await {
-            Ok(Ok(())) => {}
-            Ok(Err(e)) => return Err(e),
-            Err(_elapsed) => {
-                return Err(ExecutorError::request(StdContractError::Put {
-                    key: *key,
-                    cause: "timed out fetching related contracts".into(),
-                }));
-            }
-        }
-
-        // Merge initial_related with newly fetched states
-        let initial_owned = initial_related.clone().into_owned();
-        for (id, state) in initial_owned.states() {
-            if let Some(s) = state {
-                related_map
-                    .entry(*id)
-                    .or_insert_with(|| Some(s.clone().into_owned()));
-            }
-        }
-
-        let populated_related = RelatedContracts::from(related_map);
-        let retry_result = self
-            .runtime
-            .validate_state(key, params, state, &populated_related)
-            .map_err(|e| ExecutorError::execution(e, None))?;
-
-        if let ValidateResult::RequestRelated(_) = &retry_result {
-            return Err(ExecutorError::request(StdContractError::Put {
-                key: *key,
-                cause: "contract requested additional related contracts after first round (depth=1 limit exceeded)".into(),
-            }));
-        }
-
-        Ok(retry_result)
-    }
-
-    async fn subscribe(&mut self, key: ContractKey) -> Result<(), ExecutorError> {
-        if self.mode == OperationMode::Local {
-            return Ok(());
-        }
-        let op_manager = self
-            .op_manager
-            .as_ref()
-            .ok_or_else(|| ExecutorError::other(anyhow::anyhow!("missing op_manager")))?;
-        let executor_tx = crate::message::Transaction::new::<operations::subscribe::SubscribeMsg>();
-        // Caps total task lifetime — the inner driver's per-attempt
-        // `OPERATION_TTL = 60 s` would otherwise allow multi-attempt
-        // waits to compound. Any change here should be checked against
-        // the per-attempt budget so `MAX_RETRIES` attempts can complete
-        // within the deadline.
-        const SUBSCRIBE_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(120);
-        match tokio::time::timeout(
-            SUBSCRIBE_TIMEOUT,
-            operations::subscribe::run_executor_subscribe(
-                op_manager.clone(),
-                *key.id(),
-                executor_tx,
-            ),
-        )
-        .await
-        {
-            Ok(Ok(())) => Ok(()),
-            Ok(Err(err)) => Err(ExecutorError::other(anyhow::anyhow!("{err}"))),
-            Err(_) => Err(ExecutorError::other(anyhow::anyhow!(
-                "executor subscribe timed out after {}s",
-                SUBSCRIBE_TIMEOUT.as_secs()
-            ))),
-        }
-    }
-
-    #[inline]
-    async fn local_state_or_from_network(
-        &mut self,
-        id: &ContractInstanceId,
-        return_contract_code: bool,
-    ) -> Result<Either<WrappedState, operations::get::GetResult>, ExecutorError> {
-        // Try to get locally if we have the full key
-        if let Some(full_key) = self.lookup_key(id) {
-            if let Ok(state) = self.state_store.get(&full_key).await {
-                return Ok(Either::Left(state));
-            }
-        }
-        // Fetch from network via the sub-op GET driver. The driver
-        // delivers the resolved `GetResult` directly through a oneshot.
-        let op_manager = self
-            .op_manager
-            .as_ref()
-            .ok_or_else(|| ExecutorError::other(anyhow::anyhow!("missing op_manager")))?;
-        let (_tx, rx) =
-            operations::get::op_ctx_task::start_sub_op_get(op_manager, *id, return_contract_code);
-        // Outer callers may wrap this with a tighter budget (e.g.,
-        // `fetch_related_for_validation_network` uses
-        // `RELATED_FETCH_TIMEOUT = 10s`); when that fires first the
-        // receiver is dropped silently and the spawned sub-op task
-        // continues until OPERATION_TTL exhausts the retry loop. No
-        // leak (oneshot send-after-drop is graceful) — just a
-        // longer-lived background task.
-        const SUB_OP_FETCH_TIMEOUT: Duration = Duration::from_secs(120);
-        let outcome = tokio::time::timeout(SUB_OP_FETCH_TIMEOUT, rx)
-            .await
-            .map_err(|_| {
-                tracing::warn!(
-                    contract = %id,
-                    "sub-op GET timed out at executor"
-                );
-                ExecutorError::other(anyhow::anyhow!("sub-op GET timed out"))
-            })?
-            .map_err(|_| ExecutorError::other(anyhow::anyhow!("sub-op GET task dropped")))?;
-        match outcome {
-            operations::get::op_ctx_task::SubOpGetOutcome::Found(get_result) => {
-                Ok(Either::Right(get_result))
-            }
-            operations::get::op_ctx_task::SubOpGetOutcome::NotFound(cause) => {
-                Err(ExecutorError::other(anyhow::anyhow!(cause)))
-            }
-            operations::get::op_ctx_task::SubOpGetOutcome::Infra(err) => {
-                Err(ExecutorError::other(err))
-            }
-        }
-    }
 }
 
 /// Network-escalation half of the bridged `fetch_related_for_validation`
@@ -1522,160 +656,14 @@ pub(crate) fn set_test_network_fetch_override(stub: Option<NetworkFetchStub>) {
     TEST_NETWORK_FETCH_OVERRIDE.with(|cell| *cell.borrow_mut() = stub);
 }
 
-/// Resolve a [`MessageOrigin`] for a delegate `ApplicationMessages` request,
-/// in priority order:
-///
-/// 1. `caller_delegate` — set when another delegate dispatched this request
-///    via `OutboundDelegateMsg::SendDelegateMessage` (issue #3860). The
-///    runtime attests the caller's identity, so the receiver can authorize
-///    on it. This wins unconditionally — an inter-delegate message
-///    deliberately replaces (not composes with) any inherited WebApp origin.
-/// 2. `origin_contract` — set when a contract-backed web app dispatched
-///    this request via the WebSocket API.
-/// 3. `DELEGATE_INHERITED_ORIGINS[delegate_key]` — set when a parent
-///    delegate created this delegate via `create_delegate`, inheriting its
-///    WebApp attestation.
-///
-/// Extracted as a free function so the precedence rules can be unit-tested
-/// directly without standing up a full `Executor`.
-fn resolve_message_origin(
-    caller_delegate: Option<&DelegateKey>,
-    origin_contract: Option<&ContractInstanceId>,
-    delegate_key: &DelegateKey,
-) -> Option<MessageOrigin> {
-    if let Some(caller) = caller_delegate {
-        Some(MessageOrigin::Delegate(caller.clone()))
-    } else if let Some(contract_id) = origin_contract {
-        Some(MessageOrigin::WebApp(*contract_id))
-    } else {
-        // Plain read, no timestamp update. The "last used" time is refreshed in
-        // inbound_app_message instead, so a child that only ever gets messages
-        // from other delegates (those don't reach this branch) still counts as
-        // active and isn't dropped.
-        crate::wasm_runtime::DELEGATE_INHERITED_ORIGINS
-            .get(delegate_key)
-            .and_then(|entry| entry.origins.first().copied().map(MessageOrigin::WebApp))
-    }
-}
-
-#[cfg(test)]
-mod resolve_message_origin_tests {
-    use super::*;
-    use freenet_stdlib::prelude::CodeHash;
-
-    fn dkey(seed: u8) -> DelegateKey {
-        DelegateKey::new([seed; 32], CodeHash::new([seed; 32]))
-    }
-
-    /// Caller delegate identity wins over a concurrently-supplied WebApp
-    /// contract (regression for issue #3860 precedence rule).
-    #[test]
-    fn caller_delegate_takes_precedence_over_origin_contract() {
-        let caller = dkey(0xA1);
-        let recipient = dkey(0xB2);
-        let app_contract = ContractInstanceId::new([0xC3; 32]);
-
-        let origin = resolve_message_origin(Some(&caller), Some(&app_contract), &recipient);
-
-        match origin {
-            Some(MessageOrigin::Delegate(k)) => assert_eq!(k, caller),
-            other => panic!("Expected Delegate(caller), got {other:?}"),
-        }
-    }
-
-    /// With only `origin_contract` set, the receiver sees `WebApp(..)` —
-    /// the historical behavior for web-app-driven dispatch must be
-    /// preserved.
-    #[test]
-    fn origin_contract_alone_yields_webapp() {
-        let recipient = dkey(0xB2);
-        let app_contract = ContractInstanceId::new([0xC3; 32]);
-
-        let origin = resolve_message_origin(None, Some(&app_contract), &recipient);
-
-        match origin {
-            Some(MessageOrigin::WebApp(id)) => assert_eq!(id, app_contract),
-            other => panic!("Expected WebApp(app_contract), got {other:?}"),
-        }
-    }
-
-    /// With neither argument set and no inherited origin in the static
-    /// map, the receiver sees `None` (matches pre-#3860 behavior for
-    /// orphaned dispatches and the fall-through case for unrelated
-    /// recipients in tests).
-    #[test]
-    fn no_arguments_and_no_inherited_yields_none() {
-        // Pick a recipient key with no entry in DELEGATE_INHERITED_ORIGINS.
-        // Using a randomized seed avoids collision with anything another
-        // test populated in the same process.
-        let recipient = dkey(0xEE);
-        crate::wasm_runtime::DELEGATE_INHERITED_ORIGINS.remove(&recipient);
-
-        let origin = resolve_message_origin(None, None, &recipient);
-        assert!(origin.is_none(), "Expected None, got {origin:?}");
-    }
-
-    /// Caller delegate identity also wins over an inherited WebApp origin
-    /// in `DELEGATE_INHERITED_ORIGINS`. This documents the deliberate
-    /// "inter-delegate calls revoke inherited contract access" semantics
-    /// from the `MessageOrigin::Delegate` rustdoc.
-    #[test]
-    fn caller_delegate_overrides_inherited_origin() {
-        let caller = dkey(0xA1);
-        let recipient = dkey(0xB3);
-        let inherited_contract = ContractInstanceId::new([0xDD; 32]);
-
-        // Plant an inherited WebApp origin for the recipient so the
-        // fallback branch would have something to return.
-        crate::wasm_runtime::DELEGATE_INHERITED_ORIGINS.insert(
-            recipient.clone(),
-            crate::wasm_runtime::InheritedOriginsEntry::new(vec![inherited_contract]),
-        );
-
-        let origin = resolve_message_origin(Some(&caller), None, &recipient);
-
-        // Cleanup before assertions so a panic doesn't leak state into
-        // sibling tests sharing the same process.
-        crate::wasm_runtime::DELEGATE_INHERITED_ORIGINS.remove(&recipient);
-
-        match origin {
-            Some(MessageOrigin::Delegate(k)) => assert_eq!(k, caller),
-            other => panic!("Expected Delegate(caller), got {other:?}"),
-        }
-    }
-
-    /// Fallback branch (no live caller/origin) yields the child's inherited
-    /// WebApp origin via a pure read — it does not refresh `last_access`
-    /// (liveness lives in `inbound_app_message`). Pairs with
-    /// `no_arguments_and_no_inherited_yields_none`.
-    #[test]
-    fn inherited_origin_fallback_yields_webapp() {
-        use crate::wasm_runtime::{DELEGATE_INHERITED_ORIGINS, InheritedOriginsEntry};
-
-        let recipient = dkey(0xC5);
-        let contract = ContractInstanceId::new([0xC6; 32]);
-        DELEGATE_INHERITED_ORIGINS.insert(
-            recipient.clone(),
-            InheritedOriginsEntry::new(vec![contract]),
-        );
-
-        let origin = resolve_message_origin(None, None, &recipient);
-        DELEGATE_INHERITED_ORIGINS.remove(&recipient);
-
-        assert!(
-            matches!(origin, Some(MessageOrigin::WebApp(c)) if c == contract),
-            "fallback must yield the inherited WebApp origin, got {origin:?}"
-        );
-    }
-}
-
 #[cfg(test)]
 mod executor_pin_tests {
     /// Pin: `local_state_or_from_network` MUST use the sub-op GET
     /// driver.
     #[test]
     fn local_state_or_from_network_uses_sub_op_driver() {
-        let src = include_str!("runtime.rs");
+        // `local_state_or_from_network` lives in runtime/subscriptions.rs after the split.
+        let src = include_str!("runtime/subscriptions.rs");
         let body = src
             .split("async fn local_state_or_from_network(")
             .nth(1)
@@ -1707,7 +695,8 @@ mod executor_pin_tests {
     /// Pin: `executor::subscribe` MUST use `run_executor_subscribe`.
     #[test]
     fn executor_subscribe_uses_run_executor_subscribe() {
-        let src = include_str!("runtime.rs");
+        // `executor::subscribe` lives in runtime/subscriptions.rs after the split.
+        let src = include_str!("runtime/subscriptions.rs");
         let body = src
             .split("async fn subscribe(&mut self, key: ContractKey)")
             .nth(1)
@@ -1741,7 +730,8 @@ mod executor_pin_tests {
     /// Network-mode UPDATEs flow through `start_client_update`.
     #[test]
     fn perform_contract_update_does_not_use_network_op_request() {
-        let src = include_str!("runtime.rs");
+        // `perform_contract_update` lives in runtime/contract_ops.rs after the split.
+        let src = include_str!("runtime/contract_ops.rs");
         let body = src
             .split("async fn perform_contract_update(")
             .nth(1)
@@ -1790,11 +780,11 @@ mod executor_pin_tests {
     fn related_contract_fetch_sites_use_join_all() {
         // After the file split, `fetch_related_for_validation` lives in
         // runtime/executor_impl.rs and `fetch_related_for_validation_network`
-        // lives in runtime.rs. Search each function in its own source file so
-        // that the needle string used in the search does not appear as a
-        // string literal inside this test (which would make nth(1) extract the
-        // test code instead of the production function body).
-        const RUNTIME_SRC: &str = include_str!("runtime.rs");
+        // lives in runtime/contract_ops.rs. Search each function in its own
+        // source file so that the needle string used in the search does not
+        // appear as a string literal inside this test (which would make nth(1)
+        // extract the test code instead of the production function body).
+        const CONTRACT_OPS_SRC: &str = include_str!("runtime/contract_ops.rs");
         const EXECUTOR_IMPL_SRC: &str = include_str!("runtime/executor_impl.rs");
 
         // Each entry: (function name, file source to search, split needle).
@@ -1808,7 +798,7 @@ mod executor_pin_tests {
             ),
             (
                 "fetch_related_for_validation_network",
-                RUNTIME_SRC,
+                CONTRACT_OPS_SRC,
                 "async fn fetch_related_for_validation_network(",
             ),
         ];
@@ -2339,12 +1329,14 @@ mod state_write_attribution_pin_tests {
     //! enforcing structural invariants — see `cargo` and `rustc`'s own
     //! test suites for similar patterns).
 
-    // After the split, commit_state_write call sites live in both runtime.rs and
-    // runtime/executor_impl.rs (the generic bridged impl). Concatenate both so the
-    // count covers all chokepoints.
+    // After the split, commit_state_write call sites live in runtime.rs (the V2
+    // delegate callback installers), runtime/executor_impl.rs (the generic
+    // bridged impl), and runtime/contract_ops.rs (the concrete PUT/UPDATE
+    // chokepoints). Concatenate all three so the count covers every chokepoint.
     const RUNTIME_SRC: &str = concat!(
         include_str!("runtime.rs"),
-        include_str!("runtime/executor_impl.rs")
+        include_str!("runtime/executor_impl.rs"),
+        include_str!("runtime/contract_ops.rs")
     );
     const RING_SRC: &str = include_str!("../../ring.rs");
     const NATIVE_API_SRC: &str = include_str!("../../wasm_runtime/native_api.rs");
